Remove a coffee from the order before placing it

+++ src/main/java/org/coffeehouse/model/Coffee.java
@@ -3,7 +3,7 @@ package org.coffeehouse.model;
 import java.util.ArrayList;
 import java.util.Collection;
 
-public class Coffee extends CoffeeBase<Long> {
+public class Coffee extends CoffeeBase<Integer> {
 
     private String customerName;
     private Collection<Ingredient> extraIngredientList = new ArrayList<>();
